Move formatters into cli/ and try to fix aiohttp logs
