+++ ioccheck/cli/formatters/__init__.py
@@ -1,0 +1,8 @@
+#from ioccheck.cli.formatters.formatter import Formatter
+#from ioccheck.cli.formatters.malwarebazaar import MalwareBazaarFormatter
+#from ioccheck.cli.formatters.virustotal import VirusTotalFormatter
+
+from .formatter import Formatter
+from .malwarebazaar import MalwareBazaarFormatter
+from .virustotal import VirusTotalFormatter
+
