+++ ioccheck/cli/formatters/malwarebazaar.py
@@ -2,7 +2,7 @@ from hurry.filesize import size
 from tabulate import tabulate
 from termcolor import colored
 
-from ioccheck.formatters import Formatter
+from ioccheck.cli.formatters import Formatter
 from ioccheck.services import MalwareBazaar
