--- ioccheck/formatters/__init__.py
@@ -1,3 +1,0 @@
-from .formatter import Formatter
-from .malwarebazaar import MalwareBazaarFormatter
-from .virustotal import VirusTotalFormatter
