+++ ioccheck/ip.py
@@ -10,6 +10,10 @@ from shodan.exception import APIError
 
 logger = logging.getLogger(__name__)
 
+aiohttp_logger = logging.getLogger("aiohttp")
+aiohttp_logger.propagate = False
+aiohttp_logger.enabled = False
+
 f_handler = logging.FileHandler("ioccheck.log")
 f_handler.setLevel(logging.INFO)
