+++ ioccheck/services/malwarebazaar.py
@@ -1,10 +1,25 @@
 import json
+import logging
 
 import requests
 from backoff import expo, on_exception
 from ratelimit import RateLimitException, limits
 
 from ioccheck.services.service import Service
+
+logger = logging.getLogger(__name__)
+
+aiohttp_logger = logging.getLogger("aiohttp")
+aiohttp_logger.propagate = False
+aiohttp_logger.enabled = False
+
+f_handler = logging.FileHandler("ioccheck.log")
+f_handler.setLevel(logging.INFO)
+
+f_format = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+f_handler.setFormatter(f_format)
+
+logger.addHandler(f_handler)
 
 
 class MalwareBazaar(Service):
@@ -13,6 +28,10 @@ class MalwareBazaar(Service):
     def __init__(self, file_hash: str, api_key: str):
         self.url = "https://mb-api.abuse.ch/api/v1/"
         self.response = self._get_api_response(file_hash, api_key)
+
+        if self.response.get("query_status") == "hash_not_found":
+            return
+
         self.hashes = self._get_other_hashes(self.response)
 
         self.investigation_url = self._get_investigation_url()
