+++ ioccheck/services/virustotal.py
@@ -6,7 +6,11 @@ from ratelimit import RateLimitException, limits
 
 from ioccheck.services.service import Service
 
-logger = logging.getLogger("ioccheck")
+logger = logging.getLogger(__name__)
+
+aiohttp_logger = logging.getLogger("aiohttp")
+aiohttp_logger.propagate = False
+aiohttp_logger.enabled = False
 
 f_handler = logging.FileHandler("ioccheck.log")
 f_handler.setLevel(logging.INFO)
